+++ src/parsing/tokenization/tokenization_utils.c
@@ -72,58 +72,3 @@ char	**get_data(t_input **input, char *line, int *i)
 	}
 	return (data);
 }
-
-int	tab_len(char **tab)
-{
-	int	size;
-
-	size = 0;
-	while (tab[size])
-		size++;
-	return (size);
-}
-
-int	fill_tab_join(char **tab, char **newtab, int *word)
-{
-	int	letter;
-	int	i;
-
-	i = 0;
-	while (tab[i])
-	{
-		letter = ft_strlen(tab[i]);
-		newtab[*word] = ft_calloc(sizeof(char), letter + 1);
-		if (!newtab[*word])
-			return (1);
-		letter = 0;
-		while (tab[i][letter])
-		{
-			newtab[*word][letter] = tab[i][letter];
-			letter++;
-		}
-		newtab[*word][letter] = '\0';
-		*word += 1;
-		i++;
-	}
-	return (0);
-}
-
-char	**tab_join(char **s1, char **s2)
-{
-	char	**tab;
-	int		word;
-
-	tab = NULL;
-	word = tab_len(s1) + tab_len(s2);
-	tab = ft_calloc(sizeof(char *), (word + 1));
-	if (!tab)
-		return (NULL);
-	word = 0;
-	if (fill_tab_join(s1, tab, &word))
-		return (free_dtab(tab), free_dtab(s2), NULL);
-	if (fill_tab_join(s2, tab, &word))
-		return (free_dtab(tab), free_dtab(s2), NULL);
-	tab[word] = NULL;
-	free_dtab(s2);
-	return (tab);
-}
